chore: move useGapsLsit file to Patterns folder
